+++ Example/PaymentSheet Example/PaymentSheet Example/PaymentSheetTestPlaygroundSettings.swift
@@ -177,6 +177,7 @@ struct PaymentSheetTestPlaygroundSettings: Codable, Equatable {
         case IT
         case NG = "ng"
         case ngWallet = "ng_wallet"
+        case mondu
         case usTax = "us_tax"
         case stripeShop = "stripe_shop_test"
         case custom