+++ StripePaymentSheet/StripePaymentSheet/Source/Categories/String+Localized.swift
@@ -382,6 +382,9 @@ extension String.Localized {
             "TrueMoney mandate text. The placeholder is the merchant's name."
         )
     }
+    static var mondu_selected_message: String {
+        STPLocalizedString("Mondu selected.", "Message shown when Mondu is selected")
+    }
     static var korean_payment_method_mandate_text: String {
         STPLocalizedString(
             "By confirming your payment with NICEPAY, you allow %@ to charge this payment method for future payments in accordance with their terms.",