+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/ExpectedFormHierarchies.swift
@@ -200,6 +200,15 @@ enum ExpectedFormHierarchy {
             ])
         }
     }
+    // MARK: - Mondu
+
+    enum Mondu {
+        static var paymentIntent: FormHierarchyNode {
+            FormHierarchyNode(type: "FormElement", children: [
+                FormHierarchyNode(type: "SubtitleElement")
+            ])
+        }
+    }
     // MARK: - PAYCO
 
     enum Payco {