+++ StripePayments/StripePayments/Source/API Bindings/Models/PaymentMethods/STPPaymentMethodParams.swift
@@ -129,6 +129,8 @@ public class STPPaymentMethodParams: NSObject, STPFormEncodable {
     @objc public var scalapay: STPPaymentMethodScalapayParams?
     /// If this is a GoPay PaymentMethod, this contains additional details.
     @objc public var goPay: STPPaymentMethodGoPayParams?
+    /// If this is a Mondu PaymentMethod, this contains additional details.
+    @objc public var mondu: STPPaymentMethodMonduParams?
     /// If this is a Naira USSD PaymentMethod, this contains additional details.
     @objc public var ngUSSD: STPPaymentMethodNgUSSDParams?
     /// If this is a QRIS PaymentMethod, this contains additional details.
@@ -946,6 +948,24 @@ public class STPPaymentMethodParams: NSObject, STPFormEncodable {
         self.metadata = metadata
     }
 
+    /// Creates params for a Mondu PaymentMethod.
+    /// - Parameters:
+    ///   - mondu:       An object containing additional Mondu details.
+    ///   - billingDetails: Billing information associated with the PaymentMethod.
+    ///   - metadata:       Additional information to attach to the PaymentMethod.
+    @objc
+    public convenience init(
+        mondu: STPPaymentMethodMonduParams,
+        billingDetails: STPPaymentMethodBillingDetails?,
+        metadata: [String: String]?
+    ) {
+        self.init()
+        self.type = .mondu
+        self.mondu = mondu
+        self.billingDetails = billingDetails
+        self.metadata = metadata
+    }
+
     /// Creates params for a Naira USSD PaymentMethod.
     /// - Parameters:
     ///   - ngUSSD:       An object containing additional Naira USSD details.
@@ -1179,6 +1199,7 @@ public class STPPaymentMethodParams: NSObject, STPFormEncodable {
             NSStringFromSelector(#selector(getter: sequra)): "sequra",
             NSStringFromSelector(#selector(getter: scalapay)): "scalapay",
             NSStringFromSelector(#selector(getter: goPay)): "gopay",
+            NSStringFromSelector(#selector(getter: mondu)): "mondu",
             NSStringFromSelector(#selector(getter: ngUSSD)): "ng_ussd",
             NSStringFromSelector(#selector(getter: qris)): "qris",
             NSStringFromSelector(#selector(getter: ngWallet)): "ng_wallet",
@@ -1652,6 +1673,8 @@ extension STPPaymentMethodParams {
             scalapay = STPPaymentMethodScalapayParams()
         case .goPay:
             goPay = STPPaymentMethodGoPayParams()
+        case .mondu:
+            mondu = STPPaymentMethodMonduParams()
         case .ngUSSD:
             ngUSSD = STPPaymentMethodNgUSSDParams()
         case .qris: